% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButtR-package.R
\docType{package}
\name{ButtR-package}
\alias{ButtR}
\alias{ButtR-package}
\title{ButtR: Downloading the OzButterflies Database}
\description{
\code{ButtR} is an R package designed to simplify the download and extraction
of the \emph{OzButterflies Database}, a comprehensive dataset containing
reflectance spectra, calibrated photographs, and CO1 sequences of Australian butterflies (Reference).
}
\details{
The package includes the function \code{\link{get_Oz_butterflies}}, which allows users to download
specific subsets of the database by applying multiple filters (parameters), providing
tailored access to the data based on user-defined criteria.
}
\author{
\strong{Maintainer}: Diogo Silva \email{diogo.silva.094@ufrn.edu.br} (\href{https://orcid.org/0000-0002-0887-2001}{ORCID})

Authors:
\itemize{
  \item Jim McLean \email{jim_mclean@optusnet.com.au} (\href{https://orcid.org/0000-0001-6229-7063}{ORCID})
}

}
\keyword{internal}
