% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Oz_butterflies.R
\name{get_Oz_butterflies}
\alias{get_Oz_butterflies}
\title{Download and install the OzButterflies Database}
\usage{
get_Oz_butterflies(
  species = NULL,
  genus = NULL,
  family = NULL,
  sex = NULL,
  year = NULL,
  site = NULL,
  spectra = NULL,
  sampleIDs = NULL,
  download_images = c("raw", "jpeg"),
  download_dna = TRUE,
  save_folder = NULL,
  timeout = 10 * 60 * 60,
  quiet = FALSE,
  db_version = NA
)
}
\arguments{
\item{species}{Optional vector of binomial names of species of interest. If
specified, only species from this list will be included in the local
database.}

\item{genus}{If specified, only specimens from this genus will be installed.}

\item{family}{If specified, only specimens from this family will be
installed.}

\item{sex}{If specified, only specimens of this sex (\code{"male"},
\code{"female"} or \code{"unknown"}) will be installed.}

\item{year}{If specified, only specimens collected during these years will be
installed (options are 2022 or 2023).}

\item{site}{If specified, only specimens collected at these sites will be
installed.}

\item{spectra}{If specified, only specimens with the specified spectra value
will be installed (\code{"y"} or \code{"n"}).}

\item{sampleIDs}{If specified, only specimens with the specified IDs will be
installed.}

\item{download_images}{Specifies whether \code{"raw"} and/or \code{"jpeg"}
images should be downloaded. Only images with the specified type(s) will be
downloaded. In versions 1 to 3 of the Oz butterflies database, raw files
are in Sony raw format (\code{.ARW}) database. From version 4, raw files are in
the Adobe Digital negative format (\code{.DNG}) (see the \code{db_version} parameter).}

\item{download_dna}{If \code{TRUE} (the default), DNA files (.ab1) will be
downloaded and installed. If \code{FALSE}, DNA files will not be installed.}

\item{save_folder}{Folder where the downloaded database will be saved. This
argument must be provided by the user. If \code{save_folder = "default"},
the database will be saved in a folder called \code{"OzButterflies"} in the
current working directory.}

\item{timeout}{Maximum time allowed (in seconds) to download \emph{each} file;
default is 10 hours. The time required will depend on the speed of your
Internet connection and the parts of the database that you choose to
download. If you experience an error message such as "\code{Timeout of 36000
  seconds was reached}", try increasing the timeout.}

\item{quiet}{If \code{FALSE}, a progress bar is displayed showing the
download progress for \emph{each} file as it is downloaded from the repository,
and informational messages are printed to the console. Specify \code{quiet
  = TRUE} to prevent progress bar display.}

\item{db_version}{Version of the database to download. \code{NA} (the default)
means download the latest version. An integer version number will download
that version of the database. Note that in version \code{3} and earlier,
raw image files were in \code{.ARW} format. Starting from version 4, raw
images are in \code{.DNG} format.}
}
\value{
The installation folder (\code{save_folder}) in canonical form in
invisible form (which means it is not automatically printed).
}
\description{
Simplifies downloading the \emph{OzButterflies database} (Ref) to a local folder.
The function allows users to download specific subsets of the database by
applying multiple filters, such as species name, genus, site, family,
specific IDs, and more.
}
\details{
Be aware that downloading very large files can take many hours, depending on
the speed of your Internet connection.

\code{get_Oz_butterflies} will not remove local existing files, so subsequent
calls can be used to add to the installed database.

If you receive an intermittent error such as "\verb{status was 'SSL peer certificate or SSH remote key was not OK'}", try using a different download
method. ButtR downloads files by calling \code{\link[utils]{download.file}},
so the download method can be specified by setting the \code{download.file.method}
option; for example, \code{options(download.file.method = "curl")}. See the
\code{\link[utils]{download.file}} help for further details.
}
\examples{
\donttest{
# Download the full OzButterflies Database using the default folder name
get_Oz_butterflies(save_folder = "default")

# Get data only for Delias aganippe
get_Oz_butterflies(species = "Delias aganippe", save_folder = "Delias_aganippe") #salve_folder can be what you want

# Get data for all species of the genus Delias
get_Oz_butterflies(genus = "Delias", save_folder = "Delias_database")

# Get all species within the Nymphalidae family
get_Oz_butterflies(family = "Nymphalidae", save_folder = "Nymphalidae_data")

# Get raw files in .ARW format (from version 3 of the database),
# using the "curl" download method (Curl must be installed on your system)
options(download.file.method = "curl")
get_Oz_butterflies(
  download_images = "raw",
  db_version = 3,
  species = "Delias aganippe",
  save_folder = "Delias_raw_ARW"
)

# Get raw files in .DNG format (from version 4 of the database)
get_Oz_butterflies(
  download_images = "raw",
  db_version = 4,
  species = "Delias aganippe",
  save_folder = "Delias_raw_DNG"
)
}

}
