% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Download files from Dryad...}
\usage{
get_species(
  species = c(),
  family = NULL,
  place = c("__MACOSX", "sample frames, landmark data")
)
}
\arguments{
\item{species}{Character vector containing the names of the species to be downloaded. Default is empty.}

\item{family}{Character vector containing the names of the families from which all species will be downloaded. Default is NULL.}

\item{place}{Character vector containing the names of places to be excluded after download. Default is c("__MACOSX", "sample frames, landmark data").}
}
\description{
This function allows downloading files from Dryad for a specific butterfly species or family...
}
\examples{
\dontrun{
# Download files for a single species
get_species(species = "Delia_cristata")

# Download files for a more the one species
get_species(species = c("Delia_cristata","Delia_cumulanta"))

# Download files for an entire family
get_species(family = "Pieridae")

# Download files for species and families of a specific place
get_species(species = c("Delia_cristata", "Delia_cumulanta"),
            family = c("Pieridae", "Hesperidae"),
            place = "Brisbane")
}
}
