% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{imp}
\alias{imp}
\title{Importar arquivos TXT do repositório Dryad.}
\usage{
imp(link)
}
\arguments{
\item{dryad_link}{O link direto para o arquivo TXT hospedado no repositório Dryad.}
}
\value{
O conteúdo do arquivo TXT.
}
\description{
Esta função importa arquivos TXT de um repositório Dryad dado um link para o arquivo TXT.
}
\examples{
dryad_import("https://datadryad.org/stash/downloads/file_stream/11349")
# Retorna o conteúdo do arquivo TXT hospedado no repositório Dryad.

}
