% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{Oz_butterflies_summary}
\alias{Oz_butterflies_summary}
\title{Summarise contents of Oz butterflies database}
\usage{
Oz_butterflies_summary(db_folder = "Oz_butterflies")
}
\arguments{
\item{db_folder}{Path of folder that contains the Oz_butterflies database.}
}
\description{
Summarise contents of Oz butterflies database
}
