% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Sum two numbers.}
\usage{
somar(a, b)
}
\arguments{
\item{a}{The first number.}

\item{b}{The second number.}
}
\value{
The sum of \code{a} and \code{b}.
}
\description{
This function calculates the sum of two numbers.
}
\examples{
somar(2, 3)
# Returns: 5

somar(10, -5)
# Returns: 5

}
