% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Oz_butterflies.R
\name{get_Oz_butterflies}
\alias{get_Oz_butterflies}
\title{Download and install the OzButterflies Database}
\usage{
get_Oz_butterflies(
  species = NULL,
  genus = NULL,
  family = NULL,
  sex = NULL,
  year = NULL,
  site = NULL,
  spectra = NULL,
  sampleIDs = NULL,
  download_images = c("raw", "jpeg"),
  download_dna = TRUE,
  db_folder = "OzButterflies",
  timeout = 20 * 60,
  quiet = FALSE
)
}
\arguments{
\item{species}{Optional vector of binomial names of species of interest. If
specified, only species from this list will be included in the local
database.}

\item{genus}{If specified, only specimens from this genus will be installed.}

\item{family}{If specified, only specimens from this family will be
installed.}

\item{sex}{If specified, only specimens of this sex (\code{"male"},
\code{"female"} or \code{"unknown"}) will be installed.}

\item{year}{If specified, only specimens collected during these years will be
installed (options are 2022 or 2023).}

\item{site}{If specified, only specimens collected at these sites will be
installed.}

\item{spectra}{If specified, only specimens with the specified spectra value
will be installed (\code{"y"} or \code{"n"}).}

\item{sampleIDs}{If specified, only specimens with the specified IDs will be
installed.}

\item{download_images}{Specifies whether \code{"raw"} and/or \code{"jpeg"}
images should be downloaded. Only images with the specified type(s) will be
downloaded.}

\item{download_dna}{If \code{TRUE} (the default), DNA files (.ab1) will be
downloaded and installed. If \code{FALSE}, DNA files will not be installed.}

\item{db_folder}{Path of folder that will contain the downloaded database.}

\item{timeout}{Maximum time allowed (in seconds) to download \emph{each} file. The
time required will depend on the speed of your internet connection and the
parts of the database that you choose to download. If you experience an
error message such as "\code{Timeout of 1200 seconds was reached}", try
increasing the timeout.}

\item{quiet}{If \code{FALSE}, a progress bar is displayed showing the
download progress for \emph{each} file as it is downloaded from the
repository, and informational messages are printed to the console.}
}
\value{
The installation folder (\code{db_folder}) in canonical form in invisible
form (which means it is not automatically printed)
}
\description{
Simplifies downloading the \emph{OzButterflies database} (Ref) to a local folder.
The function allows users to download specific subsets of the database by
applying multiple filters, such as species name, genus, site, family,
specific IDs, and more.
}
\details{
\code{get_Oz_butterflies} will not remove local existing files, so subsequent
calls can be used to add to the installed database.
}
\examples{
\dontrun{
# Download the full OzButterflies Database
get_Oz_butterflies()

# Get data only for Delias aganippe
get_Oz_butterflies(species = "Delias aganippe")

# Get data for all species of the genus Delias
get_Oz_butterflies(genus = "Delias")

# Get all species within the Nymphalidae family
get_Oz_butterflies(family = "Nymphalidae")

# Get data with multiple filters
get_Oz_butterflies(species = c("Delias aganippe", "Delias mysis"))
}

}
