% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Oz_butterflies.R
\name{get_Oz_butterflies}
\alias{get_Oz_butterflies}
\title{Downloads all or part of the Oz butterflies database to a local folder}
\usage{
get_Oz_butterflies(
  species = NULL,
  genus = NULL,
  family = NULL,
  sex = NULL,
  year = NULL,
  site = NULL,
  reflectance = NULL,
  sampleIDs = NULL,
  download_images = c("raw", "jpeg"),
  db_folder = "Oz_butterflies"
)
}
\arguments{
\item{species}{Optional vector of binomial names of species of interest. If
specified, only species from this list will be included in the local
database.}

\item{genus}{If specified, only specimens from these genera will be
installed.}

\item{family}{If specified, only specimens from this family will be
installed.}

\item{sex}{If specified, only specimens of this sex (\code{"male"},
\code{"female"} or  \code{"unknown"}) will be installed.}

\item{year}{If specified, only specimens collected during these years will be
installed (options are 2022 or 2023).}

\item{site}{If specified, only specimens collected at these sites
will be installed (options are \code{"Brisbane"}, \code{"Cairns"} and
\code{"Sydney"}).}

\item{reflectance}{If specified, only specimens with the specified
reflectance will be installed (\code{"yes"} or \code{"no"}).}

\item{sampleIDs}{If specified, only specimens with the specified IDs will be
installed.}

\item{download_images}{Specifies whether \code{"raw"} and/or \code{"jpeg"}
images should be downloaded. Only images with the specified type(s) will be
downloaded.}

\item{db_folder}{Path of folder that will contain the downloaded database.}
}
\value{
Path of the downloaded folder (invisibly).
}
\description{
Simplifies downloading the Oz butterflies database to a local folder. Since
the database is quite large, download times are long and the database
requires substantial local storage space. If the entire database is not
needed, then this function saves time and local storage space by only
downloading the required parts of the database.
}
\details{
The metadata files in the database (Oz_butterflies.csv, Oz_butterflies.xslx
and Oz_butterflies.json) always describe the entire database, regardless of
whether the entire database or a subset is installed locally.
}
