% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Downloads all or part of the Oz butterflies database to a local folder.}
\usage{
get_species(
  species = NULL,
  genus = NULL,
  family = NULL,
  sex = NULL,
  year = NULL,
  location = NULL,
  reflectance = NULL,
  db_folder = "Oz_butterflies"
)
}
\arguments{
\item{species}{Optional vector of binomial names of species of interest. If
specified, only species from this list will be included in the local
database.}

\item{genus}{If specified, only specimens from these genera will be
installed.}

\item{family}{If specified, only specimens from this family will be
installed.}

\item{sex}{If specified, only specimens of this sex (\code{"male"} or \code{"female"}) will
be installed.}

\item{year}{If specified, only specimens collected during these years will be
installed (options are 2022 or 2023).}

\item{location}{If specified, only specimens collected at these locations
will be installed (options are \code{"Brisbane"}, \code{"Cairns"} and \code{"Sydney"}).}

\item{reflectance}{If specified, only specimens with the specified reflectance
will be installed (\code{"yes"} or \code{"no"}).}

\item{db_folder}{Path of folder that will contain the downloaded database.}
}
\description{
Downloads all or part of the Oz butterflies database to a local folder.
}
