% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{Oz_butterflies_summary}
\alias{Oz_butterflies_summary}
\title{Summarise contents of the OzButterflies database}
\usage{
Oz_butterflies_summary(
  db_folder = "OzButterflies",
  imgExt = "\\\\.ARW$|\\\\.arw$"
)
}
\arguments{
\item{db_folder}{Path of folder that contains the OzButterflies database.}

\item{imgExt}{Regular expression used to identify files to be counted as
images. Default is \code{.ARW} files which are the RGB and UV photos of
specimens.}
}
\value{
Data frame with 1 row and columns that summarise the database
contents. All summary statistics, apart from the \code{Images} count, describe
the entire database, regardless of whether the entire database or a subset
is installed locally.
}
\description{
Summarise contents of the OzButterflies database
}
